requirements.txt
aniso8601==10.0.0
anyio==4.8.0
asttokens==3.0.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
deep-translator==1.11.4
dnspython==2.7.0
dominate==2.9.1
dotenv==0.9.9
email_validator==2.2.0
executing==2.1.0
Faker==37.1.0
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.2.0
googletrans==4.0.2
greenlet==3.1.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
labelImg==1.8.6
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
pymongo==4.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
